Rename showStorage.ts
