+++ src/commands/insert-next-episode.ts
@@ -1,6 +1,6 @@
 
 	import { TextEditor, TextEditorEdit, window } from "vscode";
-	import ShowStorage from "../cache/anime/showStorage";
+	import ShowStorage from "../cache/anime/show-storage";
 	import LineContextFinder from "../list-parser/line-context-finder";
 	import { isEditingSimpleCursor } from "../utils/editor-utils";
 	import { MarucsAnime } from '../extension';
